Allow explicit `--include` command line arguments
The CC spec provides a `config.json` file with an array of `include_paths`,
which defines the files an engine is permitted to report issues in.

`unused` already has an `--ignore` flag for *blacklisting*, but since CC
operates on a *whitelisting* basis, it's important to allow that as well.

Munging the file-provided JSON into cmd line arguments will be done externally.